Abstract out RemoteFsTranslogWithPinnedTimestamps
